feat!: upgrade to binaryen v95 & vendor prebuilt artifacts
