+++ .gitignore
@@ -6,3 +6,4 @@ META
 .DS_Store
 *.tar.gz
 *.install
+*.zip
