--- esy.lock/opam/js_of_ocaml.3.7.1/opam
@@ -1,32 +1,0 @@
-opam-version: "2.0"
-maintainer:   "dev@ocsigen.org"
-authors:      "Ocsigen team"
-bug-reports:  "https://github.com/ocsigen/js_of_ocaml/issues"
-homepage:     "http://ocsigen.github.io/js_of_ocaml"
-dev-repo:     "git+https://github.com/ocsigen/js_of_ocaml.git"
-synopsis:     "Compiler from OCaml bytecode to Javascript"
-description: """
-Js_of_ocaml is a compiler from OCaml bytecode to JavaScript.
-It makes it possible to run pure OCaml programs in JavaScript
-environment like browsers and Node.js
-"""
-
-build: [["dune" "build" "-p" name "-j" jobs]]
-
-depends: [
-  "ocaml" {>= "4.02.0"}
-  "dune" {>= "2.5"}
-  "ocaml-migrate-parsetree" {>= "1.4" & < "2.0.0"}
-  "ppx_tools_versioned" {>= "5.2.3"}
-  "uchar"
-  "js_of_ocaml-compiler" {= version}
-]
-x-commit-hash: "888697ba6c17051c50839e35d882a8c58fdc69f5"
-url {
-  src:
-    "https://github.com/ocsigen/js_of_ocaml/releases/download/3.7.1/js_of_ocaml-3.7.1.tbz"
-  checksum: [
-    "sha256=36dba988e0ac14e54624da6f36337ad63a8d1f2b5195e7576d5e74a737ac1a44"
-    "sha512=88c01d33c8877ec9488ab8a73e84647a599120c11aa20b2ee06c02ecdf4bbbd6885e17bbf211e63299c5af244f3379f10787deb151930bfd8cb0d97fe8230156"
-  ]
-}
