+++ esy.lock/opam/js_of_ocaml.3.8.0/opam
@@ -1,0 +1,31 @@
+opam-version: "2.0"
+maintainer:   "dev@ocsigen.org"
+authors:      "Ocsigen team"
+bug-reports:  "https://github.com/ocsigen/js_of_ocaml/issues"
+homepage:     "http://ocsigen.github.io/js_of_ocaml"
+dev-repo:     "git+https://github.com/ocsigen/js_of_ocaml.git"
+synopsis:     "Compiler from OCaml bytecode to Javascript"
+description: """
+Js_of_ocaml is a compiler from OCaml bytecode to JavaScript.
+It makes it possible to run pure OCaml programs in JavaScript
+environment like browsers and Node.js
+"""
+
+build: [["dune" "build" "-p" name "-j" jobs]]
+
+depends: [
+  "ocaml" {>= "4.02.0"}
+  "dune" {>= "2.5"}
+  "ppxlib" {>= "0.15.0" }
+  "uchar"
+  "js_of_ocaml-compiler" {= version}
+]
+x-commit-hash: "09d5731241917577e9c16b6a0063c23baae00df8"
+url {
+  src:
+    "https://github.com/ocsigen/js_of_ocaml/releases/download/3.8.0/js_of_ocaml-3.8.0.tbz"
+  checksum: [
+    "sha256=9ed1424afd3eeafa5c5a031d817326edd751da58bda9a16fb4fcb1ee55f43219"
+    "sha512=e4855e242e4b0c6b396154e3d093fb5de28e4073efb1df00ee91ad52fad8530425498f4ff46631f128a9c792225f53c2046eeaea81517281cee1e3442a347578"
+  ]
+}
