--- esy.lock/opam/ocaml-migrate-parsetree.1.7.3/opam
@@ -1,37 +1,0 @@
-opam-version: "2.0"
-maintainer: "frederic.bour@lakaban.net"
-authors: [
-  "Frédéric Bour <frederic.bour@lakaban.net>"
-  "Jérémie Dimino <jeremie@dimino.org>"
-]
-license: "LGPL-2.1 with OCaml linking exception"
-homepage: "https://github.com/ocaml-ppx/ocaml-migrate-parsetree"
-bug-reports: "https://github.com/ocaml-ppx/ocaml-migrate-parsetree/issues"
-dev-repo: "git+https://github.com/ocaml-ppx/ocaml-migrate-parsetree.git"
-doc: "https://ocaml-ppx.github.io/ocaml-migrate-parsetree/"
-tags: [ "syntax" "org:ocamllabs" ]
-build: [
-  ["dune" "build" "-p" name "-j" jobs]
-]
-depends: [
-  "result"
-  "ppx_derivers"
-  "dune" {>= "1.9.0"}
-  "ocaml" {>= "4.02.3" & < "4.12"}
-]
-synopsis: "Convert OCaml parsetrees between different versions"
-description: """
-Convert OCaml parsetrees between different versions
-
-This library converts parsetrees, outcometree and ast mappers between
-different OCaml versions.  High-level functions help making PPX
-rewriters independent of a compiler version.
-"""
-url {
-  src:
-    "https://github.com/ocaml-ppx/ocaml-migrate-parsetree/releases/download/v1.7.3/ocaml-migrate-parsetree-v1.7.3.tbz"
-  checksum: [
-    "sha256=6d85717bcf476b87f290714872ed4fbde0233dc899c3158a27f439d70224fb55"
-    "sha512=fe9c74a244d160d973d8ca62e356edad4c872fc46471ddc668f854456d3979576895d446d49da2aee61c65b441b72c573225b0b254ab2eac4a0fb4debdbce9d4"
-  ]
-}
