--- esy.lock/opam/ocamlformat.0.15.0/opam
@@ -1,41 +1,0 @@
-opam-version: "2.0"
-maintainer: "OCamlFormat Team <ocamlformat-team@fb.com>"
-authors: "Josh Berdine <jjb@fb.com>"
-homepage: "https://github.com/ocaml-ppx/ocamlformat"
-bug-reports: "https://github.com/ocaml-ppx/ocamlformat/issues"
-dev-repo: "git+https://github.com/ocaml-ppx/ocamlformat.git"
-license: "MIT"
-build: [
-  ["ocaml" "tools/gen_version.mlt" "lib/Version.ml" version] {pinned}
-  ["dune" "build" "-p" name "-j" jobs]
-  ["dune" "runtest" "-p" name "-j" jobs] {with-test}
-]
-depends: [
-  "ocaml" {>= "4.06" & < "4.12"}
-  "alcotest" {with-test}
-  "base" {>= "v0.12.0" & < "v0.15"}
-  "base-unix"
-  "cmdliner"
-  "dune" {>= "2.2.0"}
-  "fix"
-  "fpath"
-  "menhir"
-  "ocaml-migrate-parsetree" {>= "1.7.3" & < "2.0.0"}
-  "ocp-indent" {with-test}
-  "odoc" {>= "1.4.2"}
-  "re"
-  "stdio" {< "v0.15"}
-  "uuseg" {>= "10.0.0"}
-  "uutf" {>= "1.0.1"}
-]
-synopsis: "Auto-formatter for OCaml code"
-description: "OCamlFormat is a tool to automatically format OCaml code in a uniform style."
-x-commit-hash: "e9e70f4d3aea202ec8ea4afd7ccb828b36822799"
-url {
-  src:
-    "https://github.com/ocaml-ppx/ocamlformat/releases/download/0.15.0/ocamlformat-0.15.0.tbz"
-  checksum: [
-    "sha256=263b4a8907e1b820ae0c1db55183ac3a44c4eaa05e8f1a144baa1a9bcadf2005"
-    "sha512=2bc0b826567fa71201031a730568b822938c66c56da47b05b8896b68439da3d0a8b9bc552fa5cacab95bf9192a1b6e583655fd90095aa686daa91cd2bdf4725a"
-  ]
-}
