+++ esy.lock/overrides/opam__s__conf_m4_opam__c__1_opam_override/package.json
@@ -1,0 +1,6 @@
+{
+  "build": "true",
+  "dependencies": {
+    "esy-m4": "esy-packages/esy-m4#c7cf0ac9221be2b1f9d90e83559ca08397a629e7"
+  }
+}
