+++ src/binaryen_stubs_modules.c
@@ -3,9 +3,10 @@
 #include <caml/alloc.h>
 #include <caml/fail.h>
 #include <caml/memory.h>
-#include <binaryen-c.h>
 
+#include "binaryen-c.h"
 #include "ocaml_helpers.h"
+
 
 /* Allocating an OCaml custom block to hold the given BinaryenModuleRef */
 static value alloc_BinaryenModuleRef(BinaryenModuleRef typ)
