+++ src/ocaml_helpers.c
@@ -6,18 +6,8 @@
 
 #include "ocaml_helpers.h"
 
-static struct custom_operations binaryen_ops = {
-  "com.binaryen.caml",
-  custom_finalize_default,
-  custom_compare_default,
-  custom_hash_default,
-  custom_serialize_default,
-  custom_deserialize_default,
-  custom_compare_ext_default,
-  custom_fixed_length_default
-};
 
-CAMLprim value 
+CAMLprim value
 caml_alloc_some(value v) {
   CAMLparam1(v);
   value some = caml_alloc_small(1, 0);
